backend/requirements.txt
aniso8601==7.0.0
Django==2.0.13
django-cors-headers==3.0.2
djangorestframework==3.9.4
graphene==2.1.8
graphene-django==2.5.0
graphql-core==2.2.1
graphql-relay==2.0.0
promise==2.2.1
psycopg2-binary==2.8.3
pytz==2019.2
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
yapf==0.28.0
